some codebase refactoring
