+++ src/lib/dom.ts
@@ -12,17 +12,11 @@ export const thumbnailProxies = <HTMLSelectElement>document.getElementById('thum
 
 export const superInput = <HTMLInputElement>document.getElementById('superInput');
 
+export const searchFilters = <HTMLSelectElement>document.getElementById('searchFilters');
+
 export const playButton = <HTMLButtonElement>document.getElementById('playButton');
 
 export const audio = <HTMLAudioElement>document.querySelector('audio');
-
-export const bitrateSelector = <HTMLSelectElement>document.getElementById('bitrateSelector');
-
-export const searchlist = <HTMLDivElement>document.getElementById('searchlist');
-
-export const suggestions = <HTMLUListElement>document.getElementById('suggestions');
-
-export const suggestionsSwitch = <HTMLSelectElement>document.getElementById('suggestionsSwitch');
 
 export const superModal = <HTMLDialogElement>document.getElementById('superModal');
