+++ src/lib/player.ts
@@ -1,16 +1,11 @@
-import { audio, bitrateSelector, discoverSwitch, favButton, favIcon, playButton, invidiousInstances } from "./dom";
+import { audio, discoverSwitch, favButton, favIcon, playButton, invidiousInstances } from "./dom";
 import { convertSStoHHMMSS, getDB, getSaved, notify, params, removeSaved, save, setMetaData } from "./utils";
 import { addListToCollection } from "../scripts/library";
 
-
 const codecSelector = <HTMLSelectElement>document.getElementById('CodecPreference');
-// set AAC as default for safari
-if (navigator.userAgent.indexOf('Safari') > -1 && navigator.userAgent.indexOf('Chrome') <= -1)
-  codecSelector.selectedIndex = 1;
+const bitrateSelector = <HTMLSelectElement>document.getElementById('bitrateSelector');
 
-const codecSaved = getSaved('codec');
-if (codecSaved)
-  codecSelector.selectedIndex = parseInt(codecSaved);
+/////////////////////////////////////////////////////////////
 
 codecSelector.addEventListener('change', async () => {
   const i = codecSelector.selectedIndex;
@@ -24,7 +19,23 @@ codecSelector.addEventListener('change', async () => {
   audio.currentTime = timeOfSwitch;
 });
 
+const codecSaved = getSaved('codec');
+if (codecSaved)
+  codecSelector.selectedIndex = parseInt(codecSaved);
+// set AAC as default for safari
+else if (navigator.userAgent.indexOf('Safari') > -1 && navigator.userAgent.indexOf('Chrome') <= -1)
+  codecSelector.selectedIndex = 1;
 
+/////////////////////////////////////////////////////////////
+
+bitrateSelector.addEventListener('change', () => {
+  const timeOfSwitch = audio.currentTime;
+  audio.src = bitrateSelector.value;
+  audio.currentTime = timeOfSwitch;
+  audio.play();
+});
+
+/////////////////////////////////////////////////////////////
 
 export default async function player(id: string | null = '') {
 
@@ -93,8 +104,8 @@ export default async function player(id: string | null = '') {
   bitrateSelector.selectedIndex = index;
   audio.src = bitrateSelector.value;
 
-
   // remove ' - Topic' from name if it exists
+
   let music = false;
   if (data.author.endsWith(' - Topic')) {
     music = true;
@@ -110,16 +121,15 @@ export default async function player(id: string | null = '') {
   );
 
 
-
   params.set('s', id);
 
   if (location.pathname === '/')
     history.replaceState({}, '', location.origin + '?s=' + params.get('s'));
 
+  const av = new URL(data.authorThumbnails[1].url);
   audio.dataset.id = id;
   audio.dataset.title = data.title;
   audio.dataset.author = data.author;
-  const av = new URL(data.authorThumbnails[1].url);
   audio.dataset.avatar = av.pathname.replace('no-rj', 'no-rw') + '?host=' + av.origin.substring(8);
   audio.dataset.duration = convertSStoHHMMSS(data.lengthSeconds);
   audio.dataset.channelUrl = data.authorUrl;
